pyproject.toml
[project]
name = "breakthevibe"
version = "0.1.0"
description = "AI-powered QA automation platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.14.0",
    "playwright>=1.48.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-xdist>=3.5.0",
    "pytest-asyncio>=0.24.0",
    "pillow>=11.0.0",
    "structlog>=24.4.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "pytest-cov>=6.0.0",
]

[project.scripts]
breakthevibe = "breakthevibe.main:cli"

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["breakthevibe/storage/migrations/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: unit tests (no I/O)",
    "integration: integration tests (DB, browser, LLM)",
]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
